pyproject.toml
[tool.poetry]
name = "learning-telegram-bot-2"
version = "0.1.0"
description = ""
authors = ["Borteq2 <borteq2@yandex.ru>"]
readme = "README.md"
packages = [{include = "learning_telegram_bot_2"}]

[tool.poetry.dependencies]
python = "^3.10"
pytelegrambotapi = "^4.10.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
